Split up application definition to look more like the other API objects
Also change the group to app.redskyops.dev
