+++ redskyctl/internal/commands/generate/experiment/groupversion_info.go
@@ -1,0 +1,28 @@
+/*
+Copyright 2020 GramLabs, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package experiment
+
+import (
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	"sigs.k8s.io/controller-runtime/pkg/scheme"
+)
+
+var (
+	GroupVersion       = schema.GroupVersion{Group: "app.redskyops.dev", Version: "v1alpha1"}
+	SchemeBuilder      = &scheme.Builder{GroupVersion: GroupVersion}
+	localSchemeBuilder = &SchemeBuilder.SchemeBuilder
+)
